cli/image_gen/requirements.txt
mlx>=0.18.0
diffusers>=0.25.0
transformers>=4.36.0
torch>=2.1.0
Pillow>=10.0.0
flask>=3.0.0
accelerate>=0.25.0
safetensors>=0.4.0
